Use np.allclose for comparison by default
At least in my use, a large number of algorithms that I compare
are numpy or scipy based and deal with floating point numbers.
Using ``numpy.equal`` for correctloness testing is not an
ideal choice for those, because round-off
errors that are unavoidalbe in any algorithm trigger an Excpetion.
Thus, I suggest to use ``numpy.allclose`` as the default comparison
function. It's going to give the same results for integer operations
and it's a more forgiving for float operations.
Testing different algorithms for their stability to round-off errors
is a separate problem.