Add `automatic_scale` option for readable timings
This option multiplies the `timings` attribute of the PerfplotData
object in-place during instantiation with a conversion factor to plot
with a more readable time-scale. Previously, if the timings were in the
order of nanoseconds then the plot y-axis would have displayed tick
labels such as 0.000000001 s. This was due to how the time-scale was
hard-coded to be in seconds. With this option enabled, the plot tick
labels will mostly be between 1-999 and the ylabel will update to
display the automatically chosen prefix i.e. milisecond (ms),
microsecond (us), or nanosecond (ns).

Note that the implemented _scale_timings() method utilizes the same
algorithm as that of the built-in timeit module to select the most
readable time scale.

Summary of Changes:
- Add option to automatically scale timings for readability
- New dictionary of plausible SI time units during benchmarking
- Add ``automatic_scale`` and ``timings_unit`` attributes/arguments
- Remove conversion of timings to seconds during plotting
- Automatically format ylabel based on selected prefix
- Change timings to be stored in a float array to all in-place mult.